Move SelectedItemProperty to base class for entry views.
