+++ src/User Interface/Views/BibliographyEditView.xaml.cs
@@ -15,42 +15,11 @@ public partial class BibliographyEditView : BibliographyPartDataGridView<Bibliog
 		InitializeComponent();
 		_mainGrid.BindingContext	= ViewModel;
 		DataGrid					= _dataGrid;
-
-		Loaded += OnLoaded;
 	}
 
 	#endregion
 
 	#region Properties
-
-	public static readonly BindableProperty SelectedItemProperty = BindableProperty.Create(
-		nameof(SelectedItem),
-		typeof(object),
-		typeof(BibliographyEditView),
-		null,
-		BindingMode.OneWayToSource);
-
-	public object? SelectedItem
-	{
-		get => GetValue(SelectedItemProperty);
-		private set => SetValue(SelectedItemProperty, value);
-	}
-
-	private void OnLoaded(object? sender, EventArgs eventArgs)
-	{
-		ViewModel.PropertyChanged += OnViewModelPropertyChanged;
-
-		SelectedItem = ViewModel.SelectedItem;
-	}
-
-	private void OnViewModelPropertyChanged(object? sender, PropertyChangedEventArgs eventArgs)
-	{
-		if (eventArgs.PropertyName == nameof(BibliographyEditViewModel.SelectedItem))
-		{
-			SelectedItem = ViewModel.SelectedItem;
-		}
-	}
-
 	#endregion
 
 	#region Button Events
